ai-analysis-service/requirements.txt
fastapi
uvicorn
python-dotenv
pillow
google-genai
agno
duckduckgo-search
